bug poco aprendizaje corregido

+++ Snake_IA_implementada/modelo.py
@@ -14,6 +14,7 @@ class AprendizajeReforzado(nn.Module):
         x = F.relu(self.linear1(x))
         x = self.linear2(x)
         return x
+
     def guardar(self, nombre_archivo='modelo.pth'):
         modelo_carpeta_ruta = './modelo'
         if not os.path.exists(modelo_carpeta_ruta):
@@ -34,7 +35,7 @@ class EntrenadorQ:
         
         estado = torch.tensor(estado, dtype=torch.float)
         estado_siguiente = torch.tensor(estado_siguiente,dtype=torch.float)
-        accion = torch.tensor(accion,dtype=torch.float)
+        accion = torch.tensor(accion,dtype=torch.long)
         recompensa = torch.tensor(recompensa,dtype=torch.float)
         # (n, x)
 
@@ -54,13 +55,13 @@ class EntrenadorQ:
             if not perder[index]:
                 nuevo_Q = recompensa[index] + self.gamma * torch.max(self.modelo(estado_siguiente[index]))
 
-            objetivo[index][torch.argmax(accion[index]).item()] = nuevo_Q 
+            objetivo[index][torch.argmax(accion[index]).item()]  = nuevo_Q 
 
         # recompensa + gamma * maximo de la (prediccion Q) 
         #clonamos la pred para tener los 3 valores
         #predicciones [argmax(accion)] = nuevo_Q
         self.optimizador.zero_grad() #funcion pytorch
-        perdida = self.criterio(objetivo,pred)
+        perdida = self.criterio(objetivo , pred)
         perdida.backward()
         self.optimizador.step()
